requirements.txt
# FastAPI 後端依賴
fastapi==0.115.0
mangum==0.19.0  # ASGI adapter for Vercel / AWS Lambda
uvicorn[standard]==0.34.0
python-multipart==0.0.16
pydantic==2.10.3
pydantic-settings==2.7.0

# 數據庫
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
asyncpg==0.30.0  # 異步 PostgreSQL 驅動

# 緩存與會話
redis==5.2.0
aioredis==2.0.1

# 身份驗證與安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.16
httpx==0.28.1
bcrypt==4.2.0

# 網絡與API
requests==2.32.3
aiohttp==3.11.10

# 數據驗證與序列化
email-validator==2.2.0
python-dateutil==2.9.0

# 配置管理
python-dotenv==1.0.1
toml==0.10.2

# 日誌記錄與監控
structlog==24.4.0
prometheus-client==0.21.2

# 健康檢查與就緒檢查
healthcheck==2.10.0

# CORS 支援
fastapi-cors==0.1.0

# 速率限制
slowapi==0.1.9
limits==3.15.0

# 現有系統整合
# 將 ontology 系統作為模組導入

# 開發工具
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
mypy==1.14.0
ruff==0.8.4
pre-commit==3.8.0

# 文檔生成
mkdocs==1.6.1
mkdocs-material==9.5.31

# 部署工具
gunicorn==21.2.0  # 生產服務器
hypercorn==0.18.1  # 替代生產服務器